Fikser en test for treffliste
